+++ Content.h
@@ -2,7 +2,7 @@
 #define CONTENT_H
 
 #include <QWidget>
-#include "core/Data.h"
+#include "core/SqlData.h"
 #include "engines/SearchEngine.h"
 #include <QLoggingCategory>
 
@@ -19,12 +19,11 @@ public:
     explicit Content(QWidget *parent = nullptr);
     ~Content();
 
-    void setData(Core::Data *data);
+    void setDatabase(Core::SqlData *db);
 
     void show();
     void hide();
     void loadTable(int id);
-    void saveTable();
     void emptyTable();
 
     int getCurrentTableId() const;
@@ -34,7 +33,7 @@ public slots:
     void tableRowsUnchecked();
     void tableRowDeleted();
     void tableRowEditingStarted(int row);
-    void tableRowEditingFinished(int row);
+    void tableRowEditingFinished(Table::row_id rowId);
     void tableRowTextEdited(int row);
     void resetTableState();
 
@@ -46,14 +45,14 @@ protected:
     virtual void paintEvent(QPaintEvent *);
 
 private:
-    void addTableRow(const QList<QString> &list);
-    void addTableEmptyRow();
+    void addTableRow(Table::row_id id, const QList<QString> &list);
+    void addTableEmptyRow(Table::row_id id);
     void createDataDirectoryIfItDoesNotExist();
     void updateTableScrollingByRow(int row);
     void adjustHorizontalScrollForCellCursor(LineEdit *cell);
 
     Ui::Content *ui;
-    Core::Data *data;
+    Core::SqlData *db;
     SearchEngine *eSearch;
     QLoggingCategory loggingCategory;
     int tableId;