+++ changelog.md
@@ -11,3 +11,4 @@
   - reimplement scroll area and scroll bar for content and menu by using CSS for styling and extending `QScrollArea` and `QScrollbar` for calculation adjustments to support all scrolling features (by mouse wheel and touchpad)
   - make table cells size adaptive based on the contents
   - add `EventCallback` internal tool for future
+  - migrate from XML files to SQLite database